microserviceA-cloud/requirements.txt

Flask==3.0.0
google-cloud-datastore==2.20.1
google-cloud-storage==2.18.2
requests==2.31.0
